+++ crates/cubecl-linalg/src/matmul/tests/mod.rs
@@ -2,9 +2,8 @@
 
 pub mod cmma_matmul;
 pub mod simple;
-mod test_macros;
+pub mod test_macros;
 pub mod test_utils;
 pub mod tiling2d;
 
-pub use test_macros::cmma::suite::*;
 pub use test_utils::*;
