+++ crates/cubecl-linalg/src/matmul/tests/test_macros/mod.rs
@@ -1,3 +1,3 @@
-pub mod cmma;
 mod simple;
+pub mod suite;
 mod tiling2d;
