+++ crates/cubecl-linalg/src/matmul/tests/test_macros/suite/precision.rs
@@ -1,0 +1,17 @@
+/// f16, f32, etc
+#[macro_export]
+macro_rules! testgen_matmul_precision {
+    ($kind: ident, $algorithm: ty) => {
+        mod f16_ty {
+            use super::*;
+
+            $crate::testgen_matmul_layout!($kind, $algorithm, (half::f16, half::f16));
+        }
+
+        // mod f32_ty {
+        //     use super::*;
+
+        //     $crate::testgen_matmul_layout!($kind, $algorithm, f32);
+        // }
+    };
+}
